feat: Complete backend reorganization to subsystem-first architecture
- Reorganize backend from functional-first to domain-driven subsystem-first structure
- Create clear subsystem boundaries: banking, monthly-budgets, project-budgets, rsu, auth, foreign-currency, onboarding, shared
- Fix all import paths across the entire codebase after reorganization
- Resolve domain boundary violations (e.g., enhanced keyword matching moved to banking domain)
- Add missing BUDGET_STATUS constants and fix import references
- Maintain git history through proper git mv commands
- Achieve 100% test success rate (37 test suites, 699 tests passing)
- Implement Domain-Driven Design principles for better maintainability and scalability

This completes the project budget system finalization with clean, enterprise-ready architecture.
