+++ backend/src/banking/services/__tests__/categoryMappingService.test.js
@@ -1,8 +1,8 @@
 const mongoose = require('mongoose');
 const categoryMappingService = require('../categoryMappingService');
 const { Category, SubCategory, Transaction } = require('../../../shared/models');
-const ManualCategorized = require('../../../shared/models/ManualCategorized');
-const { CategorizationMethod, TransactionType } = require('../../../shared/constants/enums');
+const { ManualCategorized } = require('../../../shared/models');
+const { CategorizationMethod, TransactionType } = require('../../constants/enums');
 
 describe('CategoryMappingService', () => {
   let testUserId;
