+++ backend/src/shared/models/__tests__/Tag.test.js
@@ -1,4 +1,4 @@
-const Tag = require('../Tag');
+const { Tag } = require('../');
 const User = require('../../../auth/models/User');
 
 let testUser;
