+++ backend/src/test-scenarios/patternDetectionIntegration.test.js
@@ -1,6 +1,6 @@
 const { Transaction, Category, SubCategory, User, BankAccount } = require('../shared/models');
 const TransactionPattern = require('../monthly-budgets/models/TransactionPattern');
-const { TransactionType } = require('../shared/constants/enums');
+const { TransactionType } = require('../banking/constants/enums');
 const recurrenceDetectionService = require('../monthly-budgets/services/recurrenceDetectionService');
 const budgetService = require('../monthly-budgets/services/budgetService');
